chore: update

+++ src/tju_expense/__main__.py
@@ -8,16 +8,6 @@ from dotenv import load_dotenv
 from .analyze import analyze
 from .fetch import Fetcher
 
-def setup_directories():
-    """创建必要的目录结构"""
-    data_dir = Path("data")
-    raw_dir = data_dir / "raw"
-    parsed_dir = data_dir / "parsed"
-
-    for dir_path in [data_dir, raw_dir, parsed_dir]:
-        dir_path.mkdir(exist_ok=True)
-
-    return raw_dir, parsed_dir
 
 def get_args():
     """获取命令行参数，包括Cookie值和日期范围"""
@@ -40,23 +30,28 @@ def get_args():
 
     return args
 
+
 def main():
     """主程序流程"""
     # 设置目录结构
-    raw_dir, parsed_dir = setup_directories()
+    data_dir = Path("data")
+    data_dir.mkdir(exist_ok=True)
 
     # 获取Cookie
     args = get_args()
 
-    filename = f"{args.start}_{args.end}"
-    parsed_file = parsed_dir / f"{filename}.csv"
-    fetcher = Fetcher(args.cookie)
-    user_info = fetcher.get_user_info()
-    print(user_info)
-    records = fetcher.get_records(start=args.start, end=args.end)
-    df = pd.DataFrame(records)
-    df.to_csv(parsed_file, index=False, encoding='utf-8')
-    print(f"解析后的数据已保存至: {parsed_file}")
+    parsed_file = data_dir / f"{args.start}_{args.end}.csv"
+    if parsed_file.exists():
+        print(f"已存在解析后的数据: {parsed_file}")
+    else:
+        print("正在获取数据...")
+        fetcher = Fetcher(args.cookie)
+        user_info = fetcher.get_user_info()
+        print(user_info)
+        records = fetcher.get_records(start=args.start, end=args.end)
+        df = pd.DataFrame(records)
+        df.to_csv(parsed_file, index=False, encoding='utf-8')
+        print(f"解析后的数据已保存至: {parsed_file}")
 
     analyze(parsed_file)
